ttldr: use caching to avoid the wrath of the deities
I was unaware that the maintainers of tldr prefer that clients do
not fetch individual pages, and someone made me aware of it.

Unfortunately I cannot unhear it, so here is a full caching
implementation. At least it doesn't extract the pages, and instead
only keeps a single archive, and it's still pretty small.

For posterity, the original online-only implementation is preserved
in a comment for generations to come.