tldr-fmt: refinements: "$@" + set -u, terminal width detection
The "$@" -> ${1+"$@"} is another case which should not fail but does
with pdksh and posh with empty arguments and set -u. it's a shell bug.

The terminal width change now prefers stty over tput because the latter
can "succeed" with a fake/incorrect output of 80, while stty seem to
either succeed correctly or fail.

We also now redirect /dev/tty to stdin of the command to help it if our
stdin is not a tty (e.g. while processing piped content).