tldr-fmt: fail on bad FILE arguments
Previously we could not detect bad FILE arguments except implicitly
by detecting empty input with -e1, and even then we only detected
empty if ALL of them were bad. This was not great.

Now we sort of do a dedicated-pipefail in a posix compliant way, so
that any failure with cat is detected, with or without -e1, and even
if some output was already produced.

We also manage it without tricks or perfomance impact. It's nice.