ttldr: fallback to tldr-fmt relative to [symlink] self
This allows symlinking only ttldr at PATH, which will then fallback
to tldr-fmt at the [symlink-followed] ttldr dir.

Based on readlink -f, but if it's not available then uses a fallback
which parses the output of "ls -l" (much less generic, doesn't resolve
multiple links, etc, but covers the basic case on one symlink).

The readme has been updated to mention it.