tldr-fmt: change default style for Code: bold,yellow
Previously "Code" (the actual usage example) was plain bold, which on
some systems is not easy to distinguish from not-bold.

The "Code" parts are typically the most important thing at the example
so bold-yellow is adequate.

The second-important thing is Arguments, which still default to plain
cyan, and being less prominent than "Code" is appropriate.