Revert "ttldr: invoke tldr-fmt only once"
This reverts commit 38bf556d16802eb62271e2c9e6662d8e90eccd1f.

The problem was that some unzip implementations (ahm busybox ahm)
don't error on missing file inside the zip, and while before that
commit tldr-fmt was returning error on empty input, at that commit
we relied on unzip's exit code - which fooled us.

The only case it worked was if the first attempted platform/language
combo indeed exists. If it wasn't then the page was never found.

The restored code checks empty input via tldr-fmt again, and fixes it.

Includes a trivial refactor of the code (set e=0 after tldr-fmt).